refactoring i naprawa zmiany nazwy tablicy

+++ Backend/index.js
@@ -1,6 +1,8 @@
 'use strict';
 var Hapi = require('hapi');
-
+/**************************UWAGA*********************
+ * do poprawnego działania musimy mieć plik boardsList.json
+ * z zawartością startową:             "[]"              */
 const server = new Hapi.Server({
     host: 'localhost',
     port: 3001,
@@ -9,29 +11,23 @@ const server = new Hapi.Server({
     }
 })
 
-//zwykły test
-server.route({
-    method: 'GET',
-    path: '/',
-    handler: function (request, reply) {
-        return ('It is wokring');
+function clearArray(array) {
+    while (array.length) {
+        array.pop();
     }
-});
-//pobierz wszytkie tablice
-server.route({
-    method: "GET",
-    path: "/getAllBoards",
-    handler: function (request, reply) {
+}
+
+var handlers = {
+    post: function (request, reply) {
+        return ('It is wokring');
+    },
+
+    getAllBoards: function (request, reply) {
         var fs = require("fs");
         var content = fs.readFileSync("boardsList.json");
         return (content);
-    }
-});
-//dodaj nowa tablicę
-server.route({
-    method: "POST",
-    path: "/addBoard",
-    handler: function (request, reply) {
+    },
+    addBoard: function (request, reply) {
         var fs = require("fs");
         var content = fs.readFileSync("boardsList.json");
         var as = JSON.parse(content);
@@ -46,26 +42,18 @@ server.route({
             as.push(request.payload.boardName);//dodajemy nowy element do jsona
             const jsonString = JSON.stringify(as)
             fs.writeFileSync("boardsList.json", jsonString);
-            //dodanie oddzielnego pliku na dane dla tablicy             //TODO trzeba jeszcze do jsona zapisać odpowiendie dane, bo na razie SSIE xd
-            var doPliku="{\"nazwaTablicy\": \""+request.payload.boardName+"\",\"kolumny\": []}"
+            //dodanie oddzielnego pliku na dane dla tablicy             
+            var doPliku = "{\"nazwaTablicy\": \"" + request.payload.boardName + "\",\"kolumny\": []}"
             var fs2 = require('fs');
-            fs2.writeFile(request.payload.boardName+".json", doPliku, function (err) {
+            fs2.writeFile(request.payload.boardName + ".json", doPliku, function (err) {
                 if (err) {
                     console.log(err);
                 }
             });
             return ("Dodano tablice: " + request.payload.boardName);
         }
-
-
-
-    }
-})
-//wybierz konkretną tablicę i uzysaj jej kolumny i zadania
-server.route({
-    method: "POST",
-    path: "/chooseBoard",
-    handler: function (request, reply) {
+    },
+    chooseBoard: function (request, reply) {
         var fs = require("fs");
         var content = fs.readFileSync("boardsList.json");
         //znajdź odpowidnią tablice 
@@ -78,28 +66,16 @@ server.route({
             //tablica istnieje czyli zwracamy jej kolumny
             //odczyt z pliku danych o tablicy
             var fs2 = require("fs");
-            var dataBoard = fs2.readFileSync(request.payload.boardName+".json");
-            var boardsData=JSON.parse(dataBoard);
+            var dataBoard = fs2.readFileSync(request.payload.boardName + ".json");
+            var boardsData = JSON.parse(dataBoard);
             return (boardsData);
         }
         else {
 
             return ("Podana tablica nie istnieje: " + request.payload.boardName);
         }
-
-
-    }
-})
-function clearArray(array) {
-    while (array.length) {
-      array.pop();
-    }
-  }
-//edytuj nazwe tablicy  //2 paramtery newBoardName, oldBoardName
-server.route({
-    method: "POST",
-    path: "/editBoardName",
-    handler: function (request, reply) {
+    },
+    editBoardName: function (request, reply) {
         var fs = require("fs");
         var content = fs.readFileSync("boardsList.json");
         var as = JSON.parse(content);
@@ -108,48 +84,46 @@ server.route({
             set.add(element);
         });
         if (set.has(request.payload.oldBoardName)) {
-            clearArray(as);//wyczyszczenie naszje tablicy danych
-            set.delete(request.payload.oldBoardName);
-            set.add(request.payload.newBoardName)
-            set.forEach(e=>as.push(e));//dodanie do tablic do zapisu
-            const jsonString = JSON.stringify(as)
-            fs.writeFileSync("boardsList.json", jsonString);
-            //tutaj edycja konkretnego pliku danej tablicy
-            var fs2 = require("fs");
-            var data = fs2.readFileSync(request.payload.oldBoardName+".json");
-            var as2 = JSON.parse(data);
-            as2.nazwaTablicy = request.payload.newBoardName;
-            const jsonString2 = JSON.stringify(as2)
-            var fs3 = require('fs');
-            fs3.writeFile(request.payload.newBoardName+".json", jsonString2, function (err) {
-                if (err) {
-                    console.log(err);
+            if (set.has(request.payload.newBoardName)) {
+                return("Inna tablica ma taką samę nazwę. \nZmiana nazwy nie została dokonana");
+            }
+            else {
+                clearArray(as);//wyczyszczenie naszje tablicy danych
+                set.delete(request.payload.oldBoardName);
+                set.add(request.payload.newBoardName)
+                set.forEach(e => as.push(e));//dodanie do tablic do zapisu
+                const jsonString = JSON.stringify(as)
+                fs.writeFileSync("boardsList.json", jsonString);
+                //tutaj edycja konkretnego pliku danej tablicy
+                var fs2 = require("fs");
+                var data = fs2.readFileSync(request.payload.oldBoardName + ".json");
+                var as2 = JSON.parse(data);
+                as2.nazwaTablicy = request.payload.newBoardName;
+                const jsonString2 = JSON.stringify(as2)
+                var fs3 = require('fs');
+                fs3.writeFile(request.payload.newBoardName + ".json", jsonString2, function (err) {
+                    if (err) {
+                        console.log(err);
+                    }
+                });
+                //usuń stary plik
+                try {
+                    fs2.unlinkSync(request.payload.oldBoardName + ".json")
+                    //file removed
+                } catch (err) {
+                    console.error(err)
                 }
-            });
-            //usuń stary plik
-            try {
-                fs2.unlinkSync(request.payload.oldBoardName+".json")
-                //file removed
-              } catch(err) {
-                console.error(err)
-              }
 
-            return ("Zmieniono nazwe tablicy: "+request.payload.oldBoardName+" na: "+request.payload.newBoardName);
+                return ("Zmieniono nazwe tablicy: " + request.payload.oldBoardName + " na: " + request.payload.newBoardName);
+            }
+
         }
         else {
-         
-            return ("Tablica nie istnieje: " + request.payload.oldBoardName+"\nCzyli nie zmienisz nazwy");
+
+            return ("Tablica nie istnieje: " + request.payload.oldBoardName + "\nCzyli nie zmienisz nazwy");
         }
-
-
-
-    }
-})
-//dodanie nowej kolumny //2 parametry boardName, columnName
-server.route({
-    method: "POST",
-    path: "/addColumn",
-    handler: function (request, reply) {
+    },
+    addColumn: function (request, reply) {
         var fs = require("fs");
         var content = fs.readFileSync("boardsList.json");
         //znajdź odpowidnią tablice 
@@ -162,23 +136,30 @@ server.route({
             //tablica istnieje czyli zwracamy jej kolumny
             //odczyt z pliku danych o tablicy
             var fs2 = require("fs");
-            var dataBoard = fs2.readFileSync(request.payload.boardName+".json");
+            var dataBoard = fs2.readFileSync(request.payload.boardName + ".json");
             var as2 = JSON.parse(dataBoard);
-            console.log(as2);
-            as2['kolumny'].push({ "nazwaKolumny": request.payload.columnName,"listZadan": []})
+            as2['kolumny'].push({ "nazwaKolumny": request.payload.columnName, "listZadan": [] })
             const jsonString = JSON.stringify(as2)
-            fs.writeFileSync(request.payload.boardName+".json", jsonString);
+            fs.writeFileSync(request.payload.boardName + ".json", jsonString);
             return (as2);
         }
         else {
 
             return ("Podana tablica nie istnieje: " + request.payload.boardName);
         }
-
     }
+}
 
-})
 
+
+server.route([
+    { path: '/', method: 'GET', handler: handlers.post },
+    { path: '/getAllBoards', method: 'GET', handler: handlers.getAllBoards },
+    { path: '/addBoard', method: 'POST', handler: handlers.addBoard },
+    { path: '/chooseBoard', method: 'POST', handler: handlers.chooseBoard },
+    { path: '/editBoardName', method: 'POST', handler: handlers.editBoardName },
+    { path: '/addColumn', method: 'POST', handler: handlers.addColumn }
+]);
